# backend/docs/swagger.yaml
definitions:
...
  handlers.ChangePasswordRequest:
    properties:
      new_password:
        example: newSecurePassword456
        minLength: 6
        type: string
    required:
    - new_password
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        example: invalid credentials
        type: string
    type: object
...
  handlers.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: securePassword123
        type: string
    required:
    - email
    - password
    type: object
  handlers.MessageResponse:
    properties:
      message:
        example: password changed successfully
        type: string
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: securePassword123
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  handlers.TokenResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxLCJleHAiOjE3MDI1MDAwMDB9.abcdef123456
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Изменяет пароль для авторизованного пользователя. Требуется валидный
        JWT токен в заголовке Authorization. Новый пароль должен содержать минимум
        6 символов. Старый токен остаётся валидным до истечения срока действия, рекомендуется
        получить новый токен после смены пароля.
      parameters:
      - description: Новый пароль. Минимум 6 символов, рекомендуется использовать
          буквы, цифры и специальные символы.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Пароль успешно изменён. Рекомендуется выполнить повторный вход.
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Неверный формат данных. Проверьте длину нового пароля.
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Отсутствует или невалидный JWT токен в заголовке Authorization
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Пользователь не найден (возможно был удалён)
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера при обновлении пароля
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Смена пароля текущего пользователя
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: 'Выполняет вход в систему по email и паролю. При успешной авторизации
        возвращается JWT токен, который необходимо использовать в заголовке Authorization
        для доступа к защищённым эндпоинтам. Формат: "Bearer <token>". Срок действия
        токена настраивается через переменную окружения JWT_EXPIRES_IN (по умолчанию
        24 часа).'
      parameters:
      - description: Учётные данные пользователя. Email и пароль должны совпадать
          с данными при регистрации.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная авторизация. Возвращается JWT токен для последующих
            запросов.
          schema:
            $ref: '#/definitions/handlers.TokenResponse'
        "400":
          description: Неверный формат данных в запросе
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Неверный email или пароль
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера при авторизации
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Авторизация пользователя
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя с указанным email и паролем. Email
        должен быть уникальным. Пароль должен содержать минимум 6 символов. После
        успешной регистрации возвращается JWT токен для авторизации.
      parameters:
      - description: Данные для регистрации. Email должен быть валидным, пароль минимум
          6 символов.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь успешно зарегистрирован. Возвращается JWT токен.
          schema:
            $ref: '#/definitions/handlers.TokenResponse'
        "400":
          description: Неверный формат данных. Проверьте email и длину пароля.
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Пользователь с таким email уже существует
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера при создании пользователя
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Регистрация нового пользователя
      tags:
      - auth
...
swagger: "2.0"